DDF-3945 - Separated RESTEndpoint from its implementation
