+++ catalog/rest/catalog-rest-endpoint/src/main/java/org/codice/ddf/endpoints/rest/RESTEndpoint.java
@@ -13,86 +13,26 @@
  */
 package org.codice.ddf.endpoints.rest;
 
-import static ddf.catalog.data.AttributeType.AttributeFormat.BINARY;
-import static ddf.catalog.data.AttributeType.AttributeFormat.OBJECT;
+import static org.codice.ddf.rest.service.CatalogService.BYTES;
+import static org.codice.ddf.rest.service.CatalogService.HEADER_ACCEPT_RANGES;
+import static org.codice.ddf.rest.service.CatalogService.HEADER_CONTENT_DISPOSITION;
+import static org.codice.ddf.rest.service.CatalogService.HEADER_CONTENT_LENGTH;
 
-import ddf.action.Action;
-import ddf.catalog.CatalogFramework;
-import ddf.catalog.Constants;
-import ddf.catalog.content.data.impl.ContentItemImpl;
-import ddf.catalog.content.operation.CreateStorageRequest;
-import ddf.catalog.content.operation.UpdateStorageRequest;
-import ddf.catalog.content.operation.impl.CreateStorageRequestImpl;
-import ddf.catalog.content.operation.impl.UpdateStorageRequestImpl;
-import ddf.catalog.data.Attribute;
-import ddf.catalog.data.AttributeDescriptor;
-import ddf.catalog.data.AttributeRegistry;
-import ddf.catalog.data.AttributeType;
 import ddf.catalog.data.BinaryContent;
-import ddf.catalog.data.ContentType;
 import ddf.catalog.data.Metacard;
-import ddf.catalog.data.MetacardCreationException;
-import ddf.catalog.data.MetacardType;
-import ddf.catalog.data.Result;
-import ddf.catalog.data.impl.AttributeImpl;
-import ddf.catalog.data.impl.BinaryContentImpl;
-import ddf.catalog.data.impl.MetacardImpl;
-import ddf.catalog.data.impl.MetacardTypeImpl;
-import ddf.catalog.federation.FederationException;
-import ddf.catalog.filter.FilterBuilder;
-import ddf.catalog.operation.CreateRequest;
-import ddf.catalog.operation.CreateResponse;
-import ddf.catalog.operation.QueryResponse;
-import ddf.catalog.operation.SourceInfoResponse;
-import ddf.catalog.operation.UpdateRequest;
-import ddf.catalog.operation.impl.CreateRequestImpl;
-import ddf.catalog.operation.impl.DeleteRequestImpl;
-import ddf.catalog.operation.impl.QueryImpl;
-import ddf.catalog.operation.impl.QueryRequestImpl;
-import ddf.catalog.operation.impl.SourceInfoRequestEnterprise;
-import ddf.catalog.operation.impl.UpdateRequestImpl;
 import ddf.catalog.resource.DataUsageLimitExceededException;
 import ddf.catalog.resource.Resource;
-import ddf.catalog.source.IngestException;
-import ddf.catalog.source.InternalIngestException;
-import ddf.catalog.source.SourceDescriptor;
-import ddf.catalog.source.SourceUnavailableException;
-import ddf.catalog.source.UnsupportedQueryException;
-import ddf.catalog.transform.CatalogTransformerException;
-import ddf.catalog.transform.InputTransformer;
-import ddf.mime.MimeTypeResolver;
-import ddf.mime.MimeTypeToTransformerMapper;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
 import java.io.InputStream;
-import java.io.Serializable;
 import java.io.UnsupportedEncodingException;
-import java.net.URI;
 import java.net.URLDecoder;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.time.Instant;
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
-import javax.activation.MimeType;
-import javax.activation.MimeTypeParseException;
 import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.Encoded;
 import javax.ws.rs.GET;
 import javax.ws.rs.HEAD;
+import javax.ws.rs.InternalServerErrorException;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -101,110 +41,29 @@ import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.UriBuilder;
 import javax.ws.rs.core.UriInfo;
-import net.minidev.json.JSONArray;
-import net.minidev.json.JSONObject;
-import net.minidev.json.JSONValue;
 import org.apache.commons.codec.CharEncoding;
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.io.IOUtils;
-import org.apache.commons.io.input.BoundedInputStream;
 import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang.exception.ExceptionUtils;
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.Pair;
-import org.apache.cxf.jaxrs.ext.multipart.Attachment;
 import org.apache.cxf.jaxrs.ext.multipart.MultipartBody;
-import org.codice.ddf.attachment.AttachmentInfo;
-import org.codice.ddf.attachment.AttachmentParser;
-import org.codice.ddf.platform.util.TemporaryFileBackedOutputStream;
-import org.codice.ddf.platform.util.uuidgenerator.UuidGenerator;
-import org.opengis.filter.Filter;
-import org.osgi.framework.Bundle;
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.FrameworkUtil;
-import org.osgi.framework.InvalidSyntaxException;
-import org.osgi.framework.ServiceReference;
-import org.owasp.html.HtmlPolicyBuilder;
+import org.codice.ddf.rest.service.CatalogException;
+import org.codice.ddf.rest.service.CatalogService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@Path("/")
 public class RESTEndpoint implements RESTService {
-
-  static final String DEFAULT_METACARD_TRANSFORMER = "xml";
-
-  private static final String BYTES_TO_SKIP = "BytesToSkip";
 
   private static final Logger LOGGER = LoggerFactory.getLogger(RESTEndpoint.class);
 
-  private static final Logger INGEST_LOGGER = LoggerFactory.getLogger(Constants.INGEST_LOGGER_NAME);
+  private CatalogService catalogService;
 
-  private static final String HEADER_RANGE = "Range";
-
-  private static final String HEADER_ACCEPT_RANGES = "Accept-Ranges";
-
-  private static final String HEADER_CONTENT_LENGTH = "Content-Length";
-
-  private static final String HEADER_CONTENT_DISPOSITION = "Content-Disposition";
-
-  private static final String FILE_ATTACHMENT_CONTENT_ID = "file";
-
-  private static final String FILENAME_CONTENT_DISPOSITION_PARAMETER_NAME = "filename";
-
-  private static final String BYTES = "bytes";
-
-  private static final String BYTES_EQUAL = "bytes=";
-
-  private static final String JSON_MIME_TYPE_STRING = "application/json";
-
-  public static final int MAX_INPUT_SIZE = 65_536;
-
-  private UuidGenerator uuidGenerator;
-
-  private static MimeType jsonMimeType;
-
-  static {
-    MimeType mime = null;
-    try {
-      mime = new MimeType(JSON_MIME_TYPE_STRING);
-    } catch (MimeTypeParseException e) {
-      LOGGER.info("Failed to create json mimetype.");
-    }
-    jsonMimeType = mime;
-  }
-
-  private FilterBuilder filterBuilder;
-
-  private CatalogFramework catalogFramework;
-
-  private MimeTypeToTransformerMapper mimeTypeToTransformerMapper;
-
-  private MimeTypeResolver tikaMimeTypeResolver;
-
-  private final AttachmentParser attachmentParser;
-
-  private AttributeRegistry attributeRegistry;
-
-  public RESTEndpoint(
-      CatalogFramework framework,
-      AttachmentParser attachmentParser,
-      AttributeRegistry attributeRegistry) {
+  public RESTEndpoint(CatalogService catalogService) {
     LOGGER.trace("Constructing REST Endpoint");
-    this.catalogFramework = framework;
-    this.attachmentParser = attachmentParser;
-    this.attributeRegistry = attributeRegistry;
+    this.catalogService = catalogService;
     LOGGER.trace(("Rest Endpoint constructed successfully"));
-  }
-
-  BundleContext getBundleContext() {
-    Bundle bundle = FrameworkUtil.getBundle(RESTEndpoint.class);
-    return bundle.getBundleContext();
   }
 
   @HEAD
@@ -251,120 +110,50 @@ public class RESTEndpoint implements RESTService {
       @PathParam("id") String id,
       @Context UriInfo uriInfo,
       @Context HttpServletRequest httpRequest) {
+    try {
+      Response.ResponseBuilder responseBuilder;
+      String filename = null;
 
-    Response response;
-    Response.ResponseBuilder responseBuilder;
-    QueryResponse queryResponse;
-    Metacard card = null;
+      final BinaryContent content =
+          catalogService.getHeaders(
+              sourceid, id, uriInfo.getAbsolutePath(), uriInfo.getQueryParameters());
 
-    LOGGER.trace("getHeaders");
-    URI absolutePath = uriInfo.getAbsolutePath();
-    MultivaluedMap<String, String> map = uriInfo.getQueryParameters();
-
-    if (id != null) {
-      LOGGER.debug("Got id: {}", id);
-      LOGGER.debug("Map of query parameters: \n{}", map);
-
-      Map<String, Serializable> convertedMap = convert(map);
-      convertedMap.put("url", absolutePath.toString());
-
-      LOGGER.debug("Map converted, retrieving product.");
-
-      // default to xml if no transformer specified
-      try {
-        String transformer = DEFAULT_METACARD_TRANSFORMER;
-
-        Filter filter = getFilterBuilder().attribute(Metacard.ID).is().equalTo().text(id);
-
-        Collection<String> sources = null;
-        if (sourceid != null) {
-          sources = new ArrayList<>();
-          sources.add(sourceid);
-        }
-
-        QueryRequestImpl request = new QueryRequestImpl(new QueryImpl(filter), sources);
-        request.setProperties(convertedMap);
-        queryResponse = catalogFramework.query(request, null);
-
-        // pull the metacard out of the blocking queue
-        List<Result> results = queryResponse.getResults();
-
-        // TODO: should be poll? do we want to specify a timeout? (will
-        // return null if timeout elapsed)
-        if (results != null && !results.isEmpty()) {
-          card = results.get(0).getMetacard();
-        }
-
-        if (card == null) {
-          return Response.status(Status.NOT_FOUND)
-              .entity("<pre>Unable to retrieve requested metacard.</pre>")
-              .type(MediaType.TEXT_HTML)
-              .build();
-        }
-
-        LOGGER.debug("Calling transform.");
-        final BinaryContent content = catalogFramework.transform(card, transformer, convertedMap);
-        LOGGER.debug("Read and transform complete, preparing response.");
-
-        responseBuilder = Response.noContent();
-
-        // Add the Accept-ranges header to let the client know that we accept ranges in bytes
-        responseBuilder.header(HEADER_ACCEPT_RANGES, BYTES);
-
-        String filename = null;
-
-        if (content instanceof Resource) {
-          // If we got a resource, we can extract the filename.
-          filename = ((Resource) content).getName();
-        } else {
-          String fileExtension = getFileExtensionForMimeType(content.getMimeTypeValue());
-          if (StringUtils.isNotBlank(fileExtension)) {
-            filename = id + fileExtension;
-          }
-        }
-
-        if (StringUtils.isNotBlank(filename)) {
-          LOGGER.debug("filename: {}", filename);
-          responseBuilder.header(
-              HEADER_CONTENT_DISPOSITION, "inline; filename=\"" + filename + "\"");
-        }
-
-        long size = content.getSize();
-        if (size > 0) {
-          responseBuilder.header(HEADER_CONTENT_LENGTH, size);
-        }
-
-        response = responseBuilder.build();
-
-      } catch (FederationException e) {
-        String exceptionMessage = "READ failed due to unexpected exception: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
-      } catch (CatalogTransformerException e) {
-        String exceptionMessage = "Unable to transform Metacard.  Try different transformer: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
-      } catch (SourceUnavailableException e) {
-        String exceptionMessage = "Cannot obtain query results because source is unavailable: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
-      } catch (UnsupportedQueryException e) {
-        String errorMessage = "Specified query is unsupported.  Change query and resubmit: ";
-        LOGGER.info(errorMessage, e);
-        return createBadRequestResponse(errorMessage);
-
-        // The catalog framework will throw this if any of the transformers blow up. We need to
-        // catch this exception
-        // here or else execution will return to CXF and we'll lose this message and end up with
-        // a huge stack trace
-        // in a GUI or whatever else is connected to this endpoint
-      } catch (IllegalArgumentException e) {
-        return createBadRequestResponse(e.getMessage());
+      if (content == null) {
+        return Response.status(Status.NOT_FOUND)
+            .entity("<pre>Unable to retrieve requested metacard.</pre>")
+            .type(MediaType.TEXT_HTML)
+            .build();
       }
-    } else {
-      return createBadRequestResponse("No ID specified.");
+
+      responseBuilder = Response.noContent();
+
+      // Add the Accept-ranges header to let the client know that we accept ranges in bytes
+      responseBuilder.header(HEADER_ACCEPT_RANGES, BYTES);
+
+      if (content instanceof Resource) {
+        // If we got a resource, we can extract the filename.
+        filename = ((Resource) content).getName();
+      } else {
+        String fileExtension =
+            catalogService.getFileExtensionForMimeType(content.getMimeTypeValue());
+        if (StringUtils.isNotBlank(fileExtension)) {
+          filename = id + fileExtension;
+        }
+      }
+
+      if (StringUtils.isNotBlank(filename)) {
+        LOGGER.debug("filename: {}", filename);
+        responseBuilder.header(HEADER_CONTENT_DISPOSITION, "inline; filename=\"" + filename + "\"");
+      }
+
+      long size = content.getSize();
+      if (size > 0) {
+        responseBuilder.header(HEADER_CONTENT_LENGTH, size);
+      }
+      return responseBuilder.build();
+    } catch (CatalogException e) {
+      return createBadRequestResponse(e.getMessage());
     }
-    return response;
   }
 
   /**
@@ -375,7 +164,6 @@ public class RESTEndpoint implements RESTService {
    * @param transformerParam (OPTIONAL)
    * @param uriInfo
    * @return
-   * @throws InternalServerErrorException
    */
   @GET
   @Path("/{id}")
@@ -388,14 +176,6 @@ public class RESTEndpoint implements RESTService {
     return getDocument(null, id, transformerParam, uriInfo, httpRequest);
   }
 
-  private JSONObject sourceActionToJSON(Action action) {
-    JSONObject jsonObject = new JSONObject();
-    jsonObject.put("title", action.getTitle());
-    jsonObject.put("url", action.getUrl().toString());
-    jsonObject.put("description", action.getDescription());
-    return jsonObject;
-  }
-
   /**
    * REST Get. Retrieves information regarding sources available.
    *
@@ -406,57 +186,13 @@ public class RESTEndpoint implements RESTService {
   @GET
   @Path(SOURCES_PATH)
   public Response getDocument(@Context UriInfo uriInfo, @Context HttpServletRequest httpRequest) {
-    BinaryContent content;
     ResponseBuilder responseBuilder;
-    String sourcesString;
 
-    JSONArray resultsList = new JSONArray();
-    SourceInfoResponse sources;
-    try {
-      SourceInfoRequestEnterprise sourceInfoRequestEnterprise =
-          new SourceInfoRequestEnterprise(true);
-
-      sources = catalogFramework.getSourceInfo(sourceInfoRequestEnterprise);
-      for (SourceDescriptor source : sources.getSourceInfo()) {
-        JSONObject sourceObj = new JSONObject();
-        sourceObj.put("id", source.getSourceId());
-        sourceObj.put("version", source.getVersion() != null ? source.getVersion() : "");
-        sourceObj.put("available", Boolean.valueOf(source.isAvailable()));
-
-        List<JSONObject> sourceActions =
-            source.getActions().stream().map(this::sourceActionToJSON).collect(Collectors.toList());
-
-        sourceObj.put("sourceActions", sourceActions);
-
-        JSONArray contentTypesObj = new JSONArray();
-        if (source.getContentTypes() != null) {
-          for (ContentType contentType : source.getContentTypes()) {
-            if (contentType != null && contentType.getName() != null) {
-              JSONObject contentTypeObj = new JSONObject();
-              contentTypeObj.put("name", contentType.getName());
-              contentTypeObj.put(
-                  "version", contentType.getVersion() != null ? contentType.getVersion() : "");
-              contentTypesObj.add(contentTypeObj);
-            }
-          }
-        }
-        sourceObj.put("contentTypes", contentTypesObj);
-        resultsList.add(sourceObj);
-      }
-    } catch (SourceUnavailableException e) {
-      LOGGER.info("Unable to retrieve Sources. {}", e.getMessage());
-      LOGGER.debug("Unable to retrieve Sources", e);
-    }
-
-    sourcesString = JSONValue.toJSONString(resultsList);
-    content =
-        new BinaryContentImpl(
-            new ByteArrayInputStream(sourcesString.getBytes(StandardCharsets.UTF_8)), jsonMimeType);
+    final BinaryContent content = catalogService.getDocument();
     responseBuilder = Response.ok(content.getInputStream(), content.getMimeTypeValue());
 
     // Add the Accept-ranges header to let the client know that we accept ranges in bytes
     responseBuilder.header(HEADER_ACCEPT_RANGES, BYTES);
-
     return responseBuilder.build();
   }
 
@@ -479,142 +215,71 @@ public class RESTEndpoint implements RESTService {
       @QueryParam("transform") String transformerParam,
       @Context UriInfo uriInfo,
       @Context HttpServletRequest httpRequest) {
+    try {
+      Response.ResponseBuilder responseBuilder;
+      String id = URLDecoder.decode(encodedId, CharEncoding.UTF_8);
 
-    Response response;
-    Response.ResponseBuilder responseBuilder;
-    QueryResponse queryResponse;
-    Metacard card = null;
+      final BinaryContent content =
+          catalogService.getDocument(
+              encodedSourceId,
+              encodedId,
+              transformerParam,
+              uriInfo.getAbsolutePath(),
+              uriInfo.getQueryParameters(),
+              httpRequest);
 
-    LOGGER.trace("GET");
-    URI absolutePath = uriInfo.getAbsolutePath();
-    MultivaluedMap<String, String> map = uriInfo.getQueryParameters();
-
-    if (encodedId != null) {
-      LOGGER.debug("Got id: {}", encodedId);
-      LOGGER.debug("Got service: {}", transformerParam);
-      LOGGER.debug("Map of query parameters: \n{}", map);
-
-      Map<String, Serializable> convertedMap = convert(map);
-      convertedMap.put("url", absolutePath.toString());
-
-      LOGGER.debug("Map converted, retrieving product.");
-
-      // default to xml if no transformer specified
-      try {
-        String id = URLDecoder.decode(encodedId, CharEncoding.UTF_8);
-
-        String transformer = DEFAULT_METACARD_TRANSFORMER;
-        if (transformerParam != null) {
-          transformer = transformerParam;
-        }
-        Filter filter = getFilterBuilder().attribute(Metacard.ID).is().equalTo().text(id);
-
-        Collection<String> sources = null;
-        if (encodedSourceId != null) {
-          String sourceid = URLDecoder.decode(encodedSourceId, CharEncoding.UTF_8);
-          sources = new ArrayList<>();
-          sources.add(sourceid);
-        }
-
-        QueryRequestImpl request = new QueryRequestImpl(new QueryImpl(filter), sources);
-        request.setProperties(convertedMap);
-        queryResponse = catalogFramework.query(request, null);
-
-        // pull the metacard out of the blocking queue
-        List<Result> results = queryResponse.getResults();
-
-        // TODO: should be poll? do we want to specify a timeout? (will
-        // return null if timeout elapsed)
-        if (results != null && !results.isEmpty()) {
-          card = results.get(0).getMetacard();
-        }
-
-        if (card == null) {
-          return Response.status(Status.NOT_FOUND)
-              .entity("<pre>Unable to retrieve requested metacard.</pre>")
-              .type(MediaType.TEXT_HTML)
-              .build();
-        }
-
-        // Check for Range header set the value in the map appropriately so that the
-        // catalogFramework
-        // can take care of the skipping
-        long bytesToSkip = getRangeStart(httpRequest);
-
-        if (bytesToSkip > 0) {
-          LOGGER.debug("Bytes to skip: {}", bytesToSkip);
-          convertedMap.put(BYTES_TO_SKIP, bytesToSkip);
-        }
-
-        LOGGER.debug("Calling transform.");
-        final BinaryContent content = catalogFramework.transform(card, transformer, convertedMap);
-        LOGGER.debug("Read and transform complete, preparing response.");
-
-        responseBuilder = Response.ok(content.getInputStream(), content.getMimeTypeValue());
-
-        // Add the Accept-ranges header to let the client know that we accept ranges in bytes
-        responseBuilder.header(HEADER_ACCEPT_RANGES, BYTES);
-
-        String filename = null;
-
-        if (content instanceof Resource) {
-          // If we got a resource, we can extract the filename.
-          filename = ((Resource) content).getName();
-        } else {
-          String fileExtension = getFileExtensionForMimeType(content.getMimeTypeValue());
-          if (StringUtils.isNotBlank(fileExtension)) {
-            filename = id + fileExtension;
-          }
-        }
-
-        if (StringUtils.isNotBlank(filename)) {
-          LOGGER.debug("filename: {}", filename);
-          responseBuilder.header(
-              HEADER_CONTENT_DISPOSITION, "inline; filename=\"" + filename + "\"");
-        }
-
-        long size = content.getSize();
-        if (size > 0) {
-          responseBuilder.header(HEADER_CONTENT_LENGTH, size);
-        }
-
-        response = responseBuilder.build();
-      } catch (FederationException e) {
-        String exceptionMessage = "READ failed due to unexpected exception: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
-      } catch (CatalogTransformerException e) {
-        String exceptionMessage = "Unable to transform Metacard.  Try different transformer: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
-      } catch (SourceUnavailableException e) {
-        String exceptionMessage = "Cannot obtain query results because source is unavailable: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
-      } catch (UnsupportedQueryException e) {
-        String errorMessage = "Specified query is unsupported.  Change query and resubmit: ";
-        LOGGER.info(errorMessage, e);
-        return createBadRequestResponse(errorMessage);
-      } catch (DataUsageLimitExceededException e) {
-        String errorMessage = "Unable to process request. Data usage limit exceeded: ";
-        LOGGER.debug(errorMessage, e);
-        return Response.status(Status.REQUEST_ENTITY_TOO_LARGE)
-            .entity("<pre>" + errorMessage + "</pre>")
+      if (content == null) {
+        return Response.status(Status.NOT_FOUND)
+            .entity("<pre>Unable to retrieve requested metacard.</pre>")
             .type(MediaType.TEXT_HTML)
             .build();
-        // The catalog framework will throw this if any of the transformers blow up.
-        // We need to catch this exception here or else execution will return to CXF and
-        // we'll lose this message and end up with a huge stack trace in a GUI or whatever
-        // else is connected to this endpoint
-      } catch (RuntimeException | UnsupportedEncodingException e) {
-        String exceptionMessage = "Unknown error occurred while processing request: ";
-        LOGGER.info(exceptionMessage, e);
-        throw new InternalServerErrorException(exceptionMessage);
       }
-    } else {
-      return createBadRequestResponse("No ID specified.");
+
+      LOGGER.debug("Read and transform complete, preparing response.");
+      responseBuilder = Response.ok(content.getInputStream(), content.getMimeTypeValue());
+
+      // Add the Accept-ranges header to let the client know that we accept ranges in bytes
+      responseBuilder.header(HEADER_ACCEPT_RANGES, BYTES);
+
+      String filename = null;
+
+      if (content instanceof Resource) {
+        // If we got a resource, we can extract the filename.
+        filename = ((Resource) content).getName();
+      } else {
+        String fileExtension =
+            catalogService.getFileExtensionForMimeType(content.getMimeTypeValue());
+        if (StringUtils.isNotBlank(fileExtension)) {
+          filename = id + fileExtension;
+        }
+      }
+
+      if (StringUtils.isNotBlank(filename)) {
+        LOGGER.debug("filename: {}", filename);
+        responseBuilder.header(HEADER_CONTENT_DISPOSITION, "inline; filename=\"" + filename + "\"");
+      }
+
+      long size = content.getSize();
+      if (size > 0) {
+        responseBuilder.header(HEADER_CONTENT_LENGTH, size);
+      }
+
+      return responseBuilder.build();
+
+    } catch (CatalogException e) {
+      return createBadRequestResponse(e.getMessage());
+
+    } catch (DataUsageLimitExceededException e) {
+      return Response.status(Status.REQUEST_ENTITY_TOO_LARGE)
+          .entity("<pre>" + e.getMessage() + "</pre>")
+          .type(MediaType.TEXT_HTML)
+          .build();
+
+    } catch (UnsupportedEncodingException e) {
+      String exceptionMessage = "Unknown error occurred while processing request: ";
+      LOGGER.info(exceptionMessage, e);
+      throw new InternalServerErrorException(exceptionMessage);
     }
-    return response;
   }
 
   @POST
@@ -623,76 +288,15 @@ public class RESTEndpoint implements RESTService {
       MultipartBody multipartBody,
       @Context UriInfo requestUriInfo,
       @QueryParam("transform") String transformerParam) {
-
-    LOGGER.trace("ENTERING: createMetacard");
-
-    String contentUri = multipartBody.getAttachmentObject("contentUri", String.class);
-    LOGGER.debug("contentUri = {}", contentUri);
-
-    InputStream stream = null;
-    String contentType = null;
-    Response response;
-
-    String transformer = DEFAULT_METACARD_TRANSFORMER;
-    if (transformerParam != null) {
-      transformer = transformerParam;
-    }
-
-    Attachment contentPart = multipartBody.getAttachment(FILE_ATTACHMENT_CONTENT_ID);
-    if (contentPart != null) {
-      // Example Content-Type header:
-      // Content-Type: application/json;id=geojson
-      if (contentPart.getContentType() != null) {
-        contentType = contentPart.getContentType().toString();
-      }
-
-      // Get the file contents as an InputStream and ensure the stream is positioned
-      // at the beginning
-      try {
-        stream = contentPart.getDataHandler().getInputStream();
-        if (stream != null && stream.available() == 0) {
-          stream.reset();
-        }
-      } catch (IOException e) {
-        LOGGER.info("IOException reading stream from file attachment in multipart body", e);
-      }
-    } else {
-      LOGGER.debug("No file contents attachment found");
-    }
-
-    MimeType mimeType = null;
-    if (contentType != null) {
-      try {
-        mimeType = new MimeType(contentType);
-      } catch (MimeTypeParseException e) {
-        LOGGER.debug("Unable to create MimeType from raw data {}", contentType);
-      }
-    } else {
-      LOGGER.debug("No content type specified in request");
-    }
-
     try {
-      Metacard metacard = generateMetacard(mimeType, null, stream, null);
-      String metacardId = metacard.getId();
-      LOGGER.debug("Metacard {} created", metacardId);
-      LOGGER.debug(
-          "Transforming metacard {} to {} to be able to return it to client",
-          metacardId,
-          transformer);
-      final BinaryContent content = catalogFramework.transform(metacard, transformer, null);
-      LOGGER.debug(
-          "Metacard to {} transform complete for {}, preparing response.", transformer, metacardId);
+      final BinaryContent content = catalogService.createMetacard(multipartBody, transformerParam);
 
       Response.ResponseBuilder responseBuilder =
           Response.ok(content.getInputStream(), content.getMimeTypeValue());
-      response = responseBuilder.build();
-    } catch (MetacardCreationException | CatalogTransformerException e) {
-      return createBadRequestResponse("Unable to create metacard");
+      return responseBuilder.build();
+    } catch (CatalogException e) {
+      return createBadRequestResponse(e.getMessage());
     }
-
-    LOGGER.trace("EXITING: createMetacard");
-
-    return response;
   }
 
   /**
@@ -731,64 +335,15 @@ public class RESTEndpoint implements RESTService {
       MultipartBody multipartBody,
       @QueryParam("transform") String transformerParam,
       InputStream message) {
-    LOGGER.trace("PUT");
-    Response response;
-
     try {
-      if (id != null && message != null) {
+      List<String> contentTypeList = headers.getRequestHeader(HttpHeaders.CONTENT_TYPE);
+      catalogService.updateDocument(id, contentTypeList, multipartBody, transformerParam, message);
 
-        MimeType mimeType = getMimeType(headers);
+      return Response.ok().build();
 
-        Pair<AttachmentInfo, Metacard> attachmentInfoAndMetacard = null;
-        if (multipartBody != null) {
-          List<Attachment> contentParts = multipartBody.getAllAttachments();
-          if (CollectionUtils.isNotEmpty(contentParts)) {
-            attachmentInfoAndMetacard = parseAttachments(contentParts, transformerParam);
-          } else {
-            LOGGER.debug("No file contents attachment found");
-          }
-        }
-
-        if (attachmentInfoAndMetacard == null) {
-          UpdateRequest updateRequest =
-              new UpdateRequestImpl(id, generateMetacard(mimeType, id, message, transformerParam));
-          catalogFramework.update(updateRequest);
-        } else {
-          UpdateStorageRequest streamUpdateRequest =
-              new UpdateStorageRequestImpl(
-                  Collections.singletonList(
-                      new IncomingContentItem(
-                          id,
-                          attachmentInfoAndMetacard.getLeft().getStream(),
-                          attachmentInfoAndMetacard.getLeft().getContentType(),
-                          attachmentInfoAndMetacard.getLeft().getFilename(),
-                          0,
-                          attachmentInfoAndMetacard.getRight())),
-                  null);
-          catalogFramework.update(streamUpdateRequest);
-        }
-
-        LOGGER.debug("Metacard {} updated.", id);
-        response = Response.ok().build();
-      } else {
-        String errorResponseString = "Both ID and content are needed to perform UPDATE.";
-        LOGGER.info(errorResponseString);
-        return createBadRequestResponse(errorResponseString);
-      }
-    } catch (SourceUnavailableException e) {
-      String exceptionMessage = "Cannot update catalog entry: Source is unavailable: ";
-      LOGGER.info(exceptionMessage, e);
-      throw new InternalServerErrorException(exceptionMessage);
-    } catch (InternalIngestException e) {
-      String exceptionMessage = "Error cataloging updated metadata: ";
-      LOGGER.info(exceptionMessage, e);
-      throw new InternalServerErrorException(exceptionMessage);
-    } catch (MetacardCreationException | IngestException e) {
-      String errorMessage = "Error cataloging updated metadata: ";
-      LOGGER.info(errorMessage, e);
-      return createBadRequestResponse(errorMessage);
+    } catch (CatalogException e) {
+      return createBadRequestResponse(e.getMessage());
     }
-    return response;
   }
 
   @POST
@@ -817,275 +372,20 @@ public class RESTEndpoint implements RESTService {
       MultipartBody multipartBody,
       @QueryParam("transform") String transformerParam,
       InputStream message) {
-    LOGGER.debug("POST");
-    Response response;
-
-    MimeType mimeType = getMimeType(headers);
-
     try {
-      if (message != null) {
-        Pair<AttachmentInfo, Metacard> attachmentInfoAndMetacard = null;
-        if (multipartBody != null) {
-          List<Attachment> contentParts = multipartBody.getAllAttachments();
-          if (CollectionUtils.isNotEmpty(contentParts)) {
-            attachmentInfoAndMetacard = parseAttachments(contentParts, transformerParam);
-          } else {
-            LOGGER.debug("No file contents attachment found");
-          }
-        }
+      List<String> contentTypeList = headers.getRequestHeader(HttpHeaders.CONTENT_TYPE);
+      String id =
+          catalogService.addDocument(contentTypeList, multipartBody, transformerParam, message);
 
-        CreateResponse createResponse;
-        if (attachmentInfoAndMetacard == null) {
-          CreateRequest createRequest =
-              new CreateRequestImpl(generateMetacard(mimeType, null, message, transformerParam));
-          createResponse = catalogFramework.create(createRequest);
-        } else {
-          String id =
-              attachmentInfoAndMetacard.getRight() == null
-                  ? null
-                  : attachmentInfoAndMetacard.getRight().getId();
-          if (id == null) {
-            id = uuidGenerator.generateUuid();
-          }
-          CreateStorageRequest streamCreateRequest =
-              new CreateStorageRequestImpl(
-                  Collections.singletonList(
-                      new IncomingContentItem(
-                          id,
-                          attachmentInfoAndMetacard.getLeft().getStream(),
-                          attachmentInfoAndMetacard.getLeft().getContentType(),
-                          attachmentInfoAndMetacard.getLeft().getFilename(),
-                          0L,
-                          attachmentInfoAndMetacard.getRight())),
-                  null);
-          createResponse = catalogFramework.create(streamCreateRequest);
-        }
+      UriBuilder uriBuilder = requestUriInfo.getAbsolutePathBuilder().path("/" + id);
+      ResponseBuilder responseBuilder = Response.created(uriBuilder.build());
+      responseBuilder.header(Metacard.ID, id);
 
-        String id = createResponse.getCreatedMetacards().get(0).getId();
+      return responseBuilder.build();
 
-        LOGGER.debug("Create Response id [{}]", id);
-
-        UriBuilder uriBuilder = requestUriInfo.getAbsolutePathBuilder().path("/" + id);
-
-        ResponseBuilder responseBuilder = Response.created(uriBuilder.build());
-
-        responseBuilder.header(Metacard.ID, id);
-
-        response = responseBuilder.build();
-
-        LOGGER.debug("Entry successfully saved, id: {}", id);
-        if (INGEST_LOGGER.isInfoEnabled()) {
-          INGEST_LOGGER.info("Entry successfully saved, id: {}", id);
-        }
-      } else {
-        String errorMessage = "No content found, cannot do CREATE.";
-        LOGGER.info(errorMessage);
-        return createBadRequestResponse(errorMessage);
-      }
-    } catch (SourceUnavailableException e) {
-      String exceptionMessage = "Cannot create catalog entry because source is unavailable: ";
-      LOGGER.info(exceptionMessage, e);
-      // Catalog framework logs these exceptions to the ingest logger so we don't have to.
-      throw new InternalServerErrorException(exceptionMessage);
-    } catch (InternalIngestException e) {
-      String exceptionMessage = "Error while storing entry in catalog: ";
-      LOGGER.info(exceptionMessage, e);
-      // Catalog framework logs these exceptions to the ingest logger so we don't have to.
-      throw new InternalServerErrorException(exceptionMessage);
-    } catch (MetacardCreationException | IngestException e) {
-      String errorMessage = "Error while storing entry in catalog: ";
-      LOGGER.info(errorMessage, e);
-      // Catalog framework logs these exceptions to the ingest logger so we don't have to.
-      return createBadRequestResponse(errorMessage);
-    } finally {
-      IOUtils.closeQuietly(message);
+    } catch (CatalogException e) {
+      return createBadRequestResponse(e.getMessage());
     }
-
-    return response;
-  }
-
-  Pair<AttachmentInfo, Metacard> parseAttachments(
-      List<Attachment> contentParts, String transformerParam) {
-
-    if (contentParts.size() == 1) {
-      Attachment contentPart = contentParts.get(0);
-
-      InputStream attachmentInputStream = null;
-
-      try {
-        attachmentInputStream = contentPart.getDataHandler().getInputStream();
-      } catch (IOException e) {
-        LOGGER.debug("IOException reading stream from file attachment in multipart body.", e);
-      }
-
-      return new ImmutablePair<>(
-          attachmentParser.generateAttachmentInfo(
-              attachmentInputStream,
-              contentPart.getContentType().toString(),
-              contentPart
-                  .getContentDisposition()
-                  .getParameter(FILENAME_CONTENT_DISPOSITION_PARAMETER_NAME)),
-          null);
-    }
-
-    Map<String, AttributeImpl> attributeMap = new HashMap<>();
-    Metacard metacard = null;
-    AttachmentInfo attachmentInfo = null;
-
-    for (Attachment attachment : contentParts) {
-      String name = attachment.getContentDisposition().getParameter("name");
-      String parsedName = (name.startsWith("parse.")) ? name.substring(6) : name;
-      try {
-        InputStream inputStream = attachment.getDataHandler().getInputStream();
-        if (name.equals("parse.resource")) {
-          attachmentInfo =
-              attachmentParser.generateAttachmentInfo(
-                  inputStream,
-                  attachment.getContentType().toString(),
-                  attachment
-                      .getContentDisposition()
-                      .getParameter(FILENAME_CONTENT_DISPOSITION_PARAMETER_NAME));
-        } else if (name.equals("parse.metadata")) {
-          metacard = parseMetadata(transformerParam, metacard, attachment, inputStream);
-        } else {
-          parseOverrideAttributes(attributeMap, parsedName, inputStream);
-        }
-      } catch (IOException e) {
-        LOGGER.debug(
-            "Unable to get input stream for mime attachment. Ignoring override attribute: {}",
-            name,
-            e);
-      }
-    }
-    if (attachmentInfo == null) {
-      throw new IllegalArgumentException("No parse.resource specified in request.");
-    }
-    if (metacard == null) {
-      metacard = new MetacardImpl();
-    }
-
-    Set<AttributeDescriptor> missingDescriptors = new HashSet<>();
-    for (Attribute attribute : attributeMap.values()) {
-      if (metacard.getMetacardType().getAttributeDescriptor(attribute.getName()) == null) {
-        attributeRegistry
-            .lookup(attribute.getName())
-            .ifPresent(descriptor -> missingDescriptors.add(descriptor));
-      }
-      metacard.setAttribute(attribute);
-    }
-
-    if (!missingDescriptors.isEmpty()) {
-      MetacardType original = metacard.getMetacardType();
-      MetacardImpl newMetacard = new MetacardImpl(metacard);
-      newMetacard.setType(new MetacardTypeImpl(original.getName(), original, missingDescriptors));
-      metacard = newMetacard;
-    }
-
-    return new ImmutablePair<>(attachmentInfo, metacard);
-  }
-
-  private void parseOverrideAttributes(
-      Map<String, AttributeImpl> attributeMap, String parsedName, InputStream inputStream) {
-    attributeRegistry
-        .lookup(parsedName)
-        .ifPresent(
-            descriptor ->
-                parseAttribute(
-                    attributeMap,
-                    parsedName,
-                    inputStream,
-                    descriptor.getType().getAttributeFormat()));
-  }
-
-  private void parseAttribute(
-      Map<String, AttributeImpl> attributeMap,
-      String parsedName,
-      InputStream inputStream,
-      AttributeType.AttributeFormat attributeFormat) {
-    try (InputStream is = inputStream;
-        InputStream boundedStream = new BoundedInputStream(is, MAX_INPUT_SIZE + 1)) {
-      if (attributeFormat == OBJECT) {
-        LOGGER.debug("Object type not supported for override");
-        return;
-      }
-
-      byte[] bytes = IOUtils.toByteArray(boundedStream);
-      if (bytes.length > MAX_INPUT_SIZE) {
-        LOGGER.debug("Attribute length is limited to {} bytes", MAX_INPUT_SIZE);
-        return;
-      }
-
-      AttributeImpl attribute;
-      if (attributeMap.containsKey(parsedName)) {
-        attribute = attributeMap.get(parsedName);
-      } else {
-        attribute = new AttributeImpl(parsedName, Collections.emptyList());
-        attributeMap.put(parsedName, attribute);
-      }
-
-      if (attributeFormat == BINARY) {
-        attribute.addValue(bytes);
-        return;
-      }
-
-      String value = new String(bytes, Charset.defaultCharset());
-
-      switch (attributeFormat) {
-        case XML:
-        case GEOMETRY:
-        case STRING:
-          attribute.addValue(value);
-          break;
-        case BOOLEAN:
-          attribute.addValue(Boolean.valueOf(value));
-          break;
-        case SHORT:
-          attribute.addValue(Short.valueOf(value));
-          break;
-        case LONG:
-          attribute.addValue(Long.valueOf(value));
-          break;
-        case INTEGER:
-          attribute.addValue(Integer.valueOf(value));
-          break;
-        case FLOAT:
-          attribute.addValue(Float.valueOf(value));
-          break;
-        case DOUBLE:
-          attribute.addValue(Double.valueOf(value));
-          break;
-        case DATE:
-          try {
-            Instant instant = Instant.parse(value);
-            attribute.addValue(Date.from(instant));
-          } catch (DateTimeParseException e) {
-            LOGGER.debug("Unable to parse instant '{}'", attribute, e);
-          }
-          break;
-        default:
-          LOGGER.debug("Attribute format '{}' not supported", attributeFormat);
-          break;
-      }
-    } catch (IOException e) {
-      LOGGER.debug("Unable to read attribute to override", e);
-    }
-  }
-
-  private Metacard parseMetadata(
-      String transformerParam, Metacard metacard, Attachment attachment, InputStream inputStream) {
-    String transformer = DEFAULT_METACARD_TRANSFORMER;
-    if (transformerParam != null) {
-      transformer = transformerParam;
-    }
-    try {
-      MimeType mimeType = new MimeType(attachment.getContentType().toString());
-      metacard = generateMetacard(mimeType, null, inputStream, transformer);
-    } catch (MimeTypeParseException | MetacardCreationException e) {
-      LOGGER.debug("Unable to parse metadata {}", attachment.getContentType().toString());
-    } finally {
-      IOUtils.closeQuietly(inputStream);
-    }
-    return metacard;
   }
 
   /**
@@ -1098,36 +398,13 @@ public class RESTEndpoint implements RESTService {
   @Path("/{id}")
   public Response deleteDocument(
       @PathParam("id") String id, @Context HttpServletRequest httpRequest) {
-    LOGGER.debug("DELETE");
-    Response response;
     try {
-      if (id != null) {
-        DeleteRequestImpl deleteReq =
-            new DeleteRequestImpl(new HtmlPolicyBuilder().toFactory().sanitize(id));
+      catalogService.deleteDocument(id);
+      return Response.ok(id).build();
 
-        catalogFramework.delete(deleteReq);
-        response = Response.ok(id).build();
-        LOGGER.debug("Attempting to delete Metacard with id: {}", id);
-      } else {
-        String errorMessage = "ID of entry not specified, cannot do DELETE.";
-        LOGGER.info(errorMessage);
-        return createBadRequestResponse(errorMessage);
-      }
-    } catch (SourceUnavailableException ce) {
-      String exceptionMessage =
-          "Could not delete entry from catalog since the source is unavailable: ";
-      LOGGER.info(exceptionMessage, ce);
-      throw new InternalServerErrorException(exceptionMessage);
-    } catch (InternalIngestException e) {
-      String exceptionMessage = "Error deleting entry from catalog: ";
-      LOGGER.info(exceptionMessage, e);
-      throw new InternalServerErrorException(exceptionMessage);
-    } catch (IngestException e) {
-      String errorMessage = "Error deleting entry from catalog: ";
-      LOGGER.info(errorMessage, e);
-      return createBadRequestResponse(errorMessage);
+    } catch (CatalogException e) {
+      return createBadRequestResponse(e.getMessage());
     }
-    return response;
   }
 
   private Response createBadRequestResponse(String entityMessage) {
@@ -1135,214 +412,5 @@ public class RESTEndpoint implements RESTService {
         .entity("<pre>" + entityMessage + "</pre>")
         .type(MediaType.TEXT_HTML)
         .build();
-  }
-
-  private Map<String, Serializable> convert(MultivaluedMap<String, String> map) {
-    Map<String, Serializable> convertedMap = new HashMap<>();
-
-    for (Map.Entry<String, List<String>> entry : map.entrySet()) {
-      String key = entry.getKey();
-      List<String> value = entry.getValue();
-
-      if (value.size() == 1) {
-        convertedMap.put(key, value.get(0));
-      } else {
-        // List is not serializable so we make it a String array
-        convertedMap.put(key, value.toArray());
-      }
-    }
-
-    return convertedMap;
-  }
-
-  private Metacard generateMetacard(
-      MimeType mimeType, String id, InputStream message, String transformerId)
-      throws MetacardCreationException {
-
-    Metacard generatedMetacard = null;
-
-    List<InputTransformer> listOfCandidates =
-        mimeTypeToTransformerMapper.findMatches(InputTransformer.class, mimeType);
-    List<String> stackTraceList = new ArrayList<>();
-
-    LOGGER.trace("Entering generateMetacard.");
-
-    LOGGER.debug("List of matches for mimeType [{}]: {}", mimeType, listOfCandidates);
-
-    try (TemporaryFileBackedOutputStream fileBackedOutputStream =
-        new TemporaryFileBackedOutputStream()) {
-
-      try {
-        if (null != message) {
-          IOUtils.copy(message, fileBackedOutputStream);
-        } else {
-          throw new MetacardCreationException(
-              "Could not copy bytes of content message.  Message was NULL.");
-        }
-      } catch (IOException e) {
-        throw new MetacardCreationException("Could not copy bytes of content message.", e);
-      }
-
-      Iterator<InputTransformer> it = listOfCandidates.iterator();
-      if (StringUtils.isNotEmpty(transformerId)) {
-        BundleContext bundleContext = getBundleContext();
-        Collection<ServiceReference<InputTransformer>> serviceReferences =
-            bundleContext.getServiceReferences(
-                InputTransformer.class, "(id=" + transformerId + ")");
-        it = serviceReferences.stream().map(bundleContext::getService).iterator();
-      }
-
-      while (it.hasNext()) {
-        InputTransformer transformer = it.next();
-        try (InputStream inputStreamMessageCopy =
-            fileBackedOutputStream.asByteSource().openStream()) {
-          generatedMetacard = transformer.transform(inputStreamMessageCopy);
-        } catch (CatalogTransformerException | IOException e) {
-          List<String> stackTraces = Arrays.asList(ExceptionUtils.getRootCauseStackTrace(e));
-          stackTraceList.add(
-              String.format("Transformer [%s] could not create metacard.", transformer));
-          stackTraceList.addAll(stackTraces);
-          LOGGER.debug("Transformer [{}] could not create metacard.", transformer, e);
-        }
-        if (generatedMetacard != null) {
-          break;
-        }
-      }
-
-      if (generatedMetacard == null) {
-        throw new MetacardCreationException(
-            String.format(
-                "Could not create metacard with mimeType %s : %s",
-                mimeType, StringUtils.join(stackTraceList, "\n")));
-      }
-
-      if (id != null) {
-        generatedMetacard.setAttribute(new AttributeImpl(Metacard.ID, id));
-      }
-      LOGGER.debug("Metacard id is {}", generatedMetacard.getId());
-
-    } catch (IOException e) {
-      throw new MetacardCreationException("Could not create metacard.", e);
-    } catch (InvalidSyntaxException e) {
-      throw new MetacardCreationException("Could not determine transformer", e);
-    }
-    return generatedMetacard;
-  }
-
-  private MimeType getMimeType(HttpHeaders headers) {
-    List<String> contentTypeList = headers.getRequestHeader(HttpHeaders.CONTENT_TYPE);
-
-    String singleMimeType = null;
-
-    if (contentTypeList != null && !contentTypeList.isEmpty()) {
-      singleMimeType = contentTypeList.get(0);
-      LOGGER.debug("Encountered [{}] {}", singleMimeType, HttpHeaders.CONTENT_TYPE);
-    }
-
-    MimeType mimeType = null;
-
-    // Sending a null argument to MimeType causes NPE
-    if (singleMimeType != null) {
-      try {
-        mimeType = new MimeType(singleMimeType);
-      } catch (MimeTypeParseException e) {
-        LOGGER.debug("Could not parse mime type from headers.", e);
-      }
-    }
-
-    return mimeType;
-  }
-
-  private String getFileExtensionForMimeType(String mimeType) {
-    String fileExtension = this.tikaMimeTypeResolver.getFileExtensionForMimeType(mimeType);
-    LOGGER.debug("Mime Type [{}] resolves to file extension [{}].", mimeType, fileExtension);
-    return fileExtension;
-  }
-
-  private boolean rangeHeaderExists(HttpServletRequest httpRequest) {
-    boolean response = false;
-
-    if (httpRequest != null && httpRequest.getHeader(HEADER_RANGE) != null) {
-      response = true;
-    }
-
-    return response;
-  }
-
-  // Return 0 (beginning of stream) if the range header does not exist.
-  private long getRangeStart(HttpServletRequest httpRequest) throws UnsupportedQueryException {
-    long response = 0;
-
-    if (httpRequest != null && rangeHeaderExists(httpRequest)) {
-      String rangeHeader = httpRequest.getHeader(HEADER_RANGE);
-      String range = getRange(rangeHeader);
-
-      if (range != null) {
-        response = Long.parseLong(range);
-      }
-    }
-
-    return response;
-  }
-
-  private String getRange(String rangeHeader) throws UnsupportedQueryException {
-    String response = null;
-
-    if (rangeHeader != null) {
-      if (rangeHeader.startsWith(BYTES_EQUAL)) {
-        String tempString = rangeHeader.substring(BYTES_EQUAL.length());
-        if (tempString.contains("-")) {
-          response = rangeHeader.substring(BYTES_EQUAL.length(), rangeHeader.lastIndexOf('-'));
-        } else {
-          response = rangeHeader.substring(BYTES_EQUAL.length());
-        }
-      } else {
-        throw new UnsupportedQueryException("Invalid range header: " + rangeHeader);
-      }
-    }
-
-    return response;
-  }
-
-  public void setMimeTypeToTransformerMapper(
-      MimeTypeToTransformerMapper mimeTypeToTransformerMapper) {
-    this.mimeTypeToTransformerMapper = mimeTypeToTransformerMapper;
-  }
-
-  public FilterBuilder getFilterBuilder() {
-    return filterBuilder;
-  }
-
-  public void setFilterBuilder(FilterBuilder filterBuilder) {
-    this.filterBuilder = filterBuilder;
-  }
-
-  public void setTikaMimeTypeResolver(MimeTypeResolver mimeTypeResolver) {
-    this.tikaMimeTypeResolver = mimeTypeResolver;
-  }
-
-  public void setUuidGenerator(UuidGenerator uuidGenerator) {
-    this.uuidGenerator = uuidGenerator;
-  }
-
-  protected static class IncomingContentItem extends ContentItemImpl {
-
-    private InputStream inputStream;
-
-    public IncomingContentItem(
-        String id,
-        InputStream inputStream,
-        String mimeTypeRawData,
-        String filename,
-        long size,
-        Metacard metacard) {
-      super(id, null, mimeTypeRawData, filename, size, metacard);
-      this.inputStream = inputStream;
-    }
-
-    @Override
-    public InputStream getInputStream() {
-      return inputStream;
-    }
   }
 }
