+++ catalog/rest/catalog-rest-endpoint/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -13,38 +13,15 @@
 <blueprint xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xmlns:jaxrs="http://cxf.apache.org/blueprint/jaxrs"
            xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
-		http://cxf.apache.org/blueprint/jaxrs http://cxf.apache.org/schemas/blueprint/jaxrs.xsd">
+           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
+           http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+           http://cxf.apache.org/blueprint/jaxrs
+           http://cxf.apache.org/schemas/blueprint/jaxrs.xsd">
 
-    <reference id="transformerMapper" interface="ddf.mime.MimeTypeToTransformerMapper"/>
-    <reference id="catalog" interface="ddf.catalog.CatalogFramework"/>
-    <reference id="filterBuilder" interface="ddf.catalog.filter.FilterBuilder"/>
-    <reference id="tikaMimeTypeResolver" filter="(name=tikaMimeTypeResolver)" interface="ddf.mime.MimeTypeResolver"/>
-    <reference id="uuidGenerator" interface="org.codice.ddf.platform.util.uuidgenerator.UuidGenerator" filter="(id=uuidGenerator)"/>
-    <reference id="attachmentParser" interface="org.codice.ddf.attachment.AttachmentParser"/>
-    <reference id="attributeRegistry" interface="ddf.catalog.data.AttributeRegistry"/>
-
-    <bean id="actionFactory"
-          class="org.codice.ddf.endpoints.rest.action.MetacardTransformerActionProviderFactory"/>
-
-    <reference-list id="metacardTransformerRefList"
-                    interface="ddf.catalog.transform.MetacardTransformer" availability="optional"
-                    filter="(!(generateActionProvider=false))">
-        <reference-listener bind-method="bind" unbind-method="unbind">
-            <bean class="org.codice.ddf.endpoints.rest.action.ActionProviderRegistryProxy">
-                <argument ref="actionFactory"/>
-            </bean>
-        </reference-listener>
-    </reference-list>
+    <reference id="catalogService" interface="org.codice.ddf.rest.service.CatalogService"/>
 
     <bean id="restSvc" class="org.codice.ddf.endpoints.rest.RESTEndpoint">
-        <argument ref="catalog"/>
-        <argument ref="attachmentParser"/>
-        <argument ref="attributeRegistry" />
-        <property name="filterBuilder" ref="filterBuilder"/>
-        <property name="mimeTypeToTransformerMapper" ref="transformerMapper"/>
-        <property name="tikaMimeTypeResolver" ref="tikaMimeTypeResolver"/>
-        <property name="uuidGenerator" ref="uuidGenerator" />
+        <argument ref="catalogService"/>
     </bean>
 
     <jaxrs:server id="restService" address="/catalog">
@@ -62,16 +39,5 @@
             <bean class="org.apache.cxf.jaxrs.provider.jsonp.JsonpPostStreamInterceptor"/>
         </jaxrs:outInterceptors>
     </jaxrs:server>
-
-    <bean id="viewActionProvider"
-          class="org.codice.ddf.endpoints.rest.action.ViewMetacardActionProvider">
-        <argument value="catalog.data.metacard.view"/>
-    </bean>
-
-    <service ref="viewActionProvider" interface="ddf.action.ActionProvider">
-        <service-properties>
-            <entry key="id" value="catalog.data.metacard.view"/>
-        </service-properties>
-    </service>
 
 </blueprint>
