+++ catalog/rest/catalog-rest-impl/src/main/java/org/codice/ddf/rest/impl/InternalServerErrorException.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.endpoints.rest;
+package org.codice.ddf.rest.impl;
 
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
