+++ catalog/rest/catalog-rest-impl/src/main/java/org/codice/ddf/rest/impl/action/MetacardTransformerActionProvider.java
@@ -11,10 +11,10 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.endpoints.rest.action;
+package org.codice.ddf.rest.impl.action;
 
-import static org.codice.ddf.endpoints.rest.RESTService.CONTEXT_ROOT;
-import static org.codice.ddf.endpoints.rest.RESTService.SOURCES_PATH;
+import static org.codice.ddf.rest.impl.CatalogServiceImpl.CONTEXT_ROOT;
+import static org.codice.ddf.rest.impl.CatalogServiceImpl.SOURCES_PATH;
 
 import ddf.action.Action;
 import ddf.action.impl.ActionImpl;
