+++ catalog/rest/catalog-rest-impl/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- /**
+ * Copyright (c) Codice Foundation
+ *
+ * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ *
+ **/ -->
+<blueprint xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
+           http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
+
+    <reference id="transformerMapper" interface="ddf.mime.MimeTypeToTransformerMapper"/>
+    <reference id="catalog" interface="ddf.catalog.CatalogFramework"/>
+    <reference id="filterBuilder" interface="ddf.catalog.filter.FilterBuilder"/>
+    <reference id="tikaMimeTypeResolver" filter="(name=tikaMimeTypeResolver)" interface="ddf.mime.MimeTypeResolver"/>
+    <reference id="uuidGenerator" interface="org.codice.ddf.platform.util.uuidgenerator.UuidGenerator" filter="(id=uuidGenerator)"/>
+    <reference id="attachmentParser" interface="org.codice.ddf.attachment.AttachmentParser"/>
+    <reference id="attributeRegistry" interface="ddf.catalog.data.AttributeRegistry"/>
+
+    <bean id="actionFactory"
+          class="org.codice.ddf.rest.impl.action.MetacardTransformerActionProviderFactory"/>
+
+    <reference-list id="metacardTransformerRefList"
+                    interface="ddf.catalog.transform.MetacardTransformer" availability="optional"
+                    filter="(!(generateActionProvider=false))">
+        <reference-listener bind-method="bind" unbind-method="unbind">
+            <bean class="org.codice.ddf.rest.impl.action.ActionProviderRegistryProxy">
+                <argument ref="actionFactory"/>
+            </bean>
+        </reference-listener>
+    </reference-list>
+
+    <bean id="catalogService" class="org.codice.ddf.rest.impl.CatalogServiceImpl">
+        <argument ref="catalog"/>
+        <argument ref="attachmentParser"/>
+        <argument ref="attributeRegistry" />
+        <property name="filterBuilder" ref="filterBuilder"/>
+        <property name="mimeTypeToTransformerMapper" ref="transformerMapper"/>
+        <property name="tikaMimeTypeResolver" ref="tikaMimeTypeResolver"/>
+        <property name="uuidGenerator" ref="uuidGenerator" />
+    </bean>
+
+    <service ref="catalogService" interface="org.codice.ddf.rest.service.CatalogService"/>
+
+    <bean id="viewActionProvider"
+          class="org.codice.ddf.rest.impl.action.ViewMetacardActionProvider">
+        <argument value="catalog.data.metacard.view"/>
+    </bean>
+
+    <service ref="viewActionProvider" interface="ddf.action.ActionProvider">
+        <service-properties>
+            <entry key="id" value="catalog.data.metacard.view"/>
+        </service-properties>
+    </service>
+
+</blueprint>
