+++ catalog/rest/catalog-rest-impl/src/test/java/org/codice/ddf/rest/impl/action/ViewMetacardActionProviderTest.java
@@ -11,10 +11,10 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.endpoints.rest.action;
+package org.codice.ddf.rest.impl.action;
 
-import static org.codice.ddf.endpoints.rest.RESTService.CONTEXT_ROOT;
-import static org.codice.ddf.endpoints.rest.RESTService.SOURCES_PATH;
+import static org.codice.ddf.rest.impl.CatalogServiceImpl.CONTEXT_ROOT;
+import static org.codice.ddf.rest.impl.CatalogServiceImpl.SOURCES_PATH;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
