+++ catalog/rest/pom.xml
@@ -24,6 +24,8 @@
     <name>DDF Catalog REST</name>
     <packaging>pom</packaging>
     <modules>
+        <module>catalog-rest-service</module>
+        <module>catalog-rest-impl</module>
         <module>catalog-rest-endpoint</module>
     </modules>
 </project>
